+++ packages/web3-core/CHANGELOG.md
@@ -191,4 +191,14 @@ Documentation:
 
 -   Fix the issue: "Uncaught TypeError: Class extends value undefined is not a constructor or null #6371". (#6398)
 
+## [4.3.1]
+
+### Fixed
+
+-   Fix `Web3Config` to properly update within other web3 packages when `setConfig` is used  (#6555)
+
+### Added
+
+-   Added `isMetaMaskProvider` function to check if provider is metamask (#6534)
+
 ## [Unreleased]